Closes #12
Added support to read package-info.java files and load package documentation
to insert into the generated package-index files.

Now the list of packages returned by the method generate_documents
from apidoc.py returns a dictionary instead of list containing just
the package names.

That dictionary uses the package name as a key
and each value is a a tuple in the same format used
for the documents dict returned by that same method.

Such a tuple contains 3 values: (package_name, class_name, parsed_javadoc).
For the elements of the packages dict, the only difference is that
instead of the class_name, it is repeated the package name.